examples/llamaindex/requirements.txt
llama-index>=0.11.0
llama-index-llms-openai>=0.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
